rolls buildREportTitle utility function into the reporting service.
